Renamed D2WIN_GetHwnd --> D2GFX_GetHwnd. (actually, I just removed the #define, it was needlessly pointless at best) Cleaned up some of the defines at the end of D2Ptrs.h and renamed some of the functions to more reasonable and consistent names.
Removed the hand-rolled GetUnitX/Y functions in favor of D2Client internals.


git-svn-id: https://subversion.assembla.com/svn/d2bs@1058 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ branches/patch-113/Core.h
@@ -9,4 +9,4 @@
 bool SplitLines(const std::string & str, size_t maxlen, const char delim, std::list<std::string> & lst);
 void Print(const char * szFormat, ...);
 void Say(const char* szMessage, ...);
-bool ClickMap(DWORD dwClickType, WORD wX, WORD wY, BOOL bShift, UnitAny* pUnit);
+bool ClickMap(DWORD dwClickType, int wX, int wY, BOOL bShift, UnitAny* pUnit);
